+++ src/rai_core/rai/communication/ros2/api/__init__.py
@@ -14,16 +14,15 @@
 
 from .action import ROS2ActionAPI
 from .base import IROS2Message
-from .service import ROS2ServiceAPI
-from .topic import ConfigurableROS2TopicAPI, ROS2TopicAPI, TopicConfig
-from .utils import (
+from .conversion import (
     convert_ros_img_to_base64,
     convert_ros_img_to_cv2mat,
     convert_ros_img_to_ndarray,
     import_message_from_str,
     ros2_message_to_dict,
-    wait_for_message,
 )
+from .service import ROS2ServiceAPI
+from .topic import ConfigurableROS2TopicAPI, ROS2TopicAPI, TopicConfig
 
 __all__ = [
     "ConfigurableROS2TopicAPI",
@@ -37,5 +36,4 @@ __all__ = [
     "convert_ros_img_to_ndarray",
     "import_message_from_str",
     "ros2_message_to_dict",
-    "wait_for_message",
 ]
